+++ spec/baby_squeel/nodes/node_spec.rb
@@ -8,11 +8,8 @@ describe BabySqueel::Nodes::Node do
 
   describe 'included modules' do
     subject { described_class.ancestors }
-    specify { is_expected.to include(BabySqueel::Operators::Comparison) }
     specify { is_expected.to include(BabySqueel::Operators::Equality) }
     specify { is_expected.to include(BabySqueel::Operators::Generic) }
-    specify { is_expected.to include(BabySqueel::Operators::Grouping) }
-    specify { is_expected.to include(BabySqueel::Operators::Matching) }
   end
 
   it 'extends any node with math' do