+++ spec/integration/when_having_spec.rb
@@ -4,7 +4,7 @@ describe '#when_having' do
   it 'adds a having clause' do
     relation = Post.selecting { id.count }
                    .grouping { author_id }
-                   .when_having { id.count > 5 }
+                   .when_having { id.count.gt(5) }
 
     expect(relation).to match_sql_snapshot
   end
@@ -12,7 +12,7 @@ describe '#when_having' do
   it 'adds a having clause with a calculation' do
     relation = Post.selecting { id.count }
                    .grouping { (author_id + 5 ) * 3 }
-                   .when_having { id.count > 5 }
+                   .when_having { id.count.gt(5) }
 
     expect(relation).to match_sql_snapshot
   end
@@ -21,7 +21,7 @@ describe '#when_having' do
     relation = Post.selecting { id.count }
                    .joins(:author)
                    .grouping { author.id }
-                   .when_having { author.id.count > 5 }
+                   .when_having { author.id.count.gt(5) }
 
     expect(relation).to match_sql_snapshot
   end
@@ -30,7 +30,7 @@ describe '#when_having' do
     relation = Post.selecting { author.posts.id.count }
                    .joining { author.posts }
                    .grouping { author.posts.id }
-                   .when_having { author.posts.id.count > 5 }
+                   .when_having { author.posts.id.count.gt(5) }
 
     expect(relation).to match_sql_snapshot(variants: ['8.1'])
   end